requirements.txt
aniso8601==9.0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
attrs==22.1.0
backports.zoneinfo==0.2.1
bidict==0.22.0
cachelib==0.10.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==7.1.2
coverage==7.3.2
distlib==0.3.6
Django==4.2.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==1.16.0
drf-yasg==1.21.7
eventlet==0.30.0
filelock==3.8.0
Flask==1.1.2
Flask-CAS==1.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-Moment==1.0.5
Flask-RESTful==0.3.9
Flask-Session==0.3.2
Flask-SocketIO==4.3.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.3
gunicorn==20.0.4
httplib2==0.20.4
idna==3.4
importlib-metadata==4.12.0
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
oauth2client==4.1.3
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pymongo==4.3.3
pyparsing==3.0.9
pytest==7.1.3
python-engineio==3.14.2
python-socketio==4.6.1
pytz==2022.4
PyYAML==6.0.1
requests==2.28.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.3.8
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.5
waitress==2.1.2
Werkzeug==0.15.5
WTForms==3.0.1
xmltodict==0.13.0
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
